requirements.txt
Backtesting==0.3.3
binance-client==1.0.0
numpy==1.21.4
pandas==1.3.5
python-binance==1.0.15
tqdm==4.62.3